selftests: ublk: add utils.h
